fix(gcp): Proper transform load GCP IAM (#1896)
### Summary
Follow standard convention of get load transform in the GCP IAM module



### Related issues or links
> Include links to relevant issues or other pages.

- https://github.com/cartography-cncf/cartography/issues/1723
(technically this doesn't do much for this issue because IAM was already
using the data model and is scoping things to projects, but I took a
look at IAM because of this issue)


### Checklist
Provide proof that this works (this makes reviews move faster). Please
perform one or more of the following:
- [x] Update/add unit or integration tests.
- [ ] Include a screenshot showing what the graph looked like before and
after your changes.
- [ ] Include console log trace showing what happened before and after
your changes.

---------

Signed-off-by: Kunaal Sikka <kunaal@subimage.io>