fix: aioboto3 to aiobotocore, change build to dev Dockerfile (#1966)
### Summary
The `build-docker-image` test runs on every commit with the goal of
flagging whether a change will break the Docker image that is pushed to
GHCR.

_However_, this test actually just pulls the latest Cartography from
Pypi. It doesn't take into account the changes in the given PR. We now
change this by building the dev Dockerfile that uses the code in the
given branch.


### Related issues or links

- https://github.com/cartography-cncf/cartography/issues/1963


### Checklist

Provide proof that this works (this makes reviews move faster). Please
perform one or more of the following:
- [ ] Update/add unit or integration tests.
- [ ] Include a screenshot showing what the graph looked like before and
after your changes.
- [ ] Include console log trace showing what happened before and after
your changes.

If you are changing a node or relationship:
- [ ] Update the
[schema](https://github.com/cartography-cncf/cartography/tree/master/docs/root/modules)
and
[readme](https://github.com/cartography-cncf/cartography/blob/master/docs/schema/README.md).

If you are implementing a new intel module:
- [ ] Use the NodeSchema [data
model](https://cartography-cncf.github.io/cartography/dev/writing-intel-modules.html#defining-a-node).

- [ ] Confirm that the linter actually passes (submitting a PR where the
linter fails shows reviewers that you did not test your code and will
delay your review).

---------

Signed-off-by: Kunaal Sikka <kunaal@subimage.io>