fix: missing link between AWSUser and AWSGroup (#2035)
### Summary
> Describe your changes.

Fixes #2032. Includes a test that was previously missing.

### Related issues or links
> Include links to relevant issues or other pages.

#2032.


### Checklist

Provide proof that this works (this makes reviews move faster). Please
perform one or more of the following:
- [x] Update/add unit or integration tests.
- [ ] Include a screenshot showing what the graph looked like before and
after your changes.
- [ ] Include console log trace showing what happened before and after
your changes.

If you are changing a node or relationship:
- [ ] Update the
[schema](https://github.com/cartography-cncf/cartography/tree/master/docs/root/modules)
and
[readme](https://github.com/cartography-cncf/cartography/blob/master/docs/schema/README.md).

- [x] Confirm that the linter actually passes (submitting a PR where the
linter fails shows reviewers that you did not test your code and will
delay your review).

Signed-off-by: Alex Chantavy <alex@subimage.io>