fix(aws): Add region field to AWS Identity Center and AWS Permission Set Nodes (#2111)
### Summary
> Describe your changes.
- This PR just adds the region field to the data models for both
identity center and permission sets so that field appears on the
respective nodes.
- The load function for both modules already passes in the region as a
kwarg so we can just use that to populate the region field without
having to modify the intel modules themeselves.


### Related issues or links
> Include links to relevant issues or other pages.

- https://github.com/cartography-cncf/cartography/issues/2109


### Checklist

Provide proof that this works (this makes reviews move faster). Please
perform one or more of the following:
- [x] Include a screenshot showing what the graph looked like before and
after your changes.
<img width="1826" height="1139" alt="permsetic"
src="https://github.com/user-attachments/assets/b74480ab-8178-4f2e-9a40-460616c4f471"
/>


If you are changing a node or relationship:
- [x] Update the
[schema](https://github.com/cartography-cncf/cartography/tree/master/docs/root/modules)
and
[readme](https://github.com/cartography-cncf/cartography/blob/master/docs/schema/README.md).

---------

Signed-off-by: shyammukund <shyam@subimage.io>