fix (AWS): Set RoleHint to include region for PermissionSets/AWS Roles not in us-east-1 (#2114)
### Summary
> The (:AWSPermissionSet)-[:ASSIGNED_TO_ROLE]-(:AWSRole) is drawn by
match the role hint (basically an arn) of the permission set to the arn
of the role. AWS Role arns are unique in that if the role in in the
region us-east-1 the region won't show up in the arn but if it is in any
other region that region will show in the arn. In cases where we want to
see this rel for roles outside of us-east-1 the matching logic fails
because the role hint on the permission set node never contained the
region while the roles arn does.

> This PR adds a change where a permission set in a region that isn't us
east 1 the role hint will contain the region so that it can properly
match with its associated role.



### Related issues or links
> Include links to relevant issues or other pages.

-https://github.com/cartography-cncf/cartography/issues/2110


### Checklist

Provide proof that this works (this makes reviews move faster). Please
perform one or more of the following:
- [x] Update/add unit or integration tests.

Signed-off-by: shyammukund <shyam@subimage.io>
Co-authored-by: Alex Chantavy <alex@subimage.io>